% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phydose-documentData.r
\docType{data}
\name{AML38}
\alias{AML38}
\title{phydose tutorial dataset of a patient (AML-38) with acute myeloid leukemia}
\format{
a list of four items:
\describe{
 \item{trees}{a list of 316 candidate set binary tree matrices with clones as rows and mutations as columns}
 \item{fmatrices}{a  frequency matrix with samples as the rows and mutation frequencies as the columns}
 \item{u_list}{a list of 316 clonal prevalence matrices with samples as the rows and clones as the columns}
 \item{graph}{a list of 316 dot format strings for each respective tree in the candate set}
}
}
\source{
\url{https://www.biorxiv.org/content/10.1101/2020.02.07.925743v1.abstract}
}
\usage{
AML38
}
\description{
phydose tutorial dataset of a patient (AML-38) with acute myeloid leukemia
}
\keyword{datasets}
