% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phydose-documentData.r
\docType{data}
\name{AML38DFF}
\alias{AML38DFF}
\title{AML38 corresponding distinguishing feature families}
\format{
a list of distinguishing feature families each containing a list of distinguishing f
eatures comprised of vectors of featurettes
}
\usage{
AML38DFF
}
\description{
AML38 corresponding distinguishing feature families
}
\keyword{datasets}
