% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phydose-plotting.r
\name{CreateDOT}
\alias{CreateDOT}
\title{Convert a binary tree to dot format}
\usage{
CreateDOT(tree)
}
\arguments{
\item{tree}{a binary matrix}
}
\value{
a string representing the string in dot format
}
\description{
Convert a binary tree to dot format
}
