% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phydose-readdata.r
\name{.CleanFmatrix}
\alias{.CleanFmatrix}
\title{Cleans the raw dataframe frequency matrix and converts it into a matrix with samples as rows and mutations as columns}
\usage{
.CleanFmatrix(df)
}
\arguments{
\item{df}{the frequency dataframe read in from the input file}
}
\value{
a matrix representing the frequency matrix with samples as rows and columns as mutations
}
\description{
Cleans the raw dataframe frequency matrix and converts it into a matrix with samples as rows and mutations as columns
}
