% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phydose-readdata.r
\name{.CreateGraphDot}
\alias{.CreateGraphDot}
\title{Create a tree in dot format}
\usage{
.CreateGraphDot(vec)
}
\arguments{
\item{vec}{character vector with the edge list}
}
\value{
a character vector representing the tree in dot format
}
\description{
Create a tree in dot format
}
