% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phydose-readdata.r
\name{.CreateKey}
\alias{.CreateKey}
\title{Create a key that maps the columns of the trees and frequencies matrices to 
the mutations}
\usage{
.CreateKey(mat)
}
\arguments{
\item{mat}{an input matrix where the column names are used as the key}
}
\value{
a dataframe creating the mapping of column ids to mutations
}
\description{
Create a key that maps the columns of the trees and frequencies matrices to 
the mutations
}
