% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phydose-readdata.r
\name{.CreateParentList}
\alias{.CreateParentList}
\title{Creates a list of of the parents for each node}
\usage{
.CreateParentList(vecs)
}
\arguments{
\item{vecs}{the portion of the file that contains the edge list of the tree}
}
\value{
a list of vectors for each node that contains the names of all ancestors of that node
}
\description{
Creates a list of of the parents for each node
}
