% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phydose-readdata.r
\name{.CreateTree}
\alias{.CreateTree}
\title{Creates a binary matrix tree with clones as rows and mutations as columns from a vector of string edge lists}
\usage{
.CreateTree(vecs)
}
\arguments{
\item{vecs}{a vector of strings with a space separator where the left string represents the parent and the 
right string represents the child}
}
\value{
a binary matrix with clones as rows and mutations as columns
}
\description{
Creates a binary matrix tree with clones as rows and mutations as columns from a vector of string edge lists
}
