% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phydose-readdata.r
\name{.GetAllF}
\alias{.GetAllF}
\title{Converts all frequency matrices in the character vector to a list of frequency matrices}
\usage{
.GetAllF(charVec)
}
\arguments{
\item{charVec}{a character vector that contains all of the text from the input file
that has trees and frequency matrices alternating}
}
\value{
a list of frequency matrices
}
\description{
Converts all frequency matrices in the character vector to a list of frequency matrices
}
