% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phydose-readdata.r
\name{.GetAllTrees}
\alias{.GetAllTrees}
\title{Helper function to read in all trees from a file and convert to binary matrices}
\usage{
.GetAllTrees(charVec)
}
\arguments{
\item{charVec}{a character vector with all lines of the file containing the trees
and frequency matrices alternating}
}
\value{
a list of trees in the format of a binary matrix with clones as rows and mutations as columns
}
\description{
Helper function to read in all trees from a file and convert to binary matrices
}
