% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phydose-readdata.r
\name{.SortMat}
\alias{.SortMat}
\title{sort a matrix that represents a tree}
\usage{
.SortMat(mat)
}
\arguments{
\item{mat}{a binary matrix}
}
\value{
a matrix where the columns are sorted by descending number of 1's an the rows
and rows are sorted ascending order of total 1's
}
\description{
sort a matrix that represents a tree
}
