% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phydose-mainfunctions.R
\name{generateDistFeat}
\alias{generateDistFeat}
\title{Generate the distinguishing feature family for a set of trees}
\usage{
generateDistFeat(treeList)
}
\arguments{
\item{treeList}{a list of trees as binary matrices}
}
\value{
a list of lists of featurettes.
}
\description{
For each tree in the input list, generateDistFeat finds the corresponding distinguishing feature family which
is a set of distinguishing features made up of root to mutation paths (featurettes) that unqiuely distinguishes that tree from 
all other trees in the input list.
}
\examples{
generateDistFeat(phydata$trees)
\dontrun{generateDistFeat(AML38$trees)}

}
