% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phydose-documentData.r
\docType{data}
\name{phydata}
\alias{phydata}
\title{phydose example dataset of two trees}
\format{
a list of four items:
\describe{
 \item{trees}{a list of a candidate set binary tree matrices with clones as rows and mutations as columns}
 \item{fmatrices}{a list of frequency matrices with samples as the rows and mutation frequencies as the columns}
 \item{u_list}{a list of clonal prevalence matrices with samples as the rows and clones as the columns}
 \item{graph}{a list of dot format string for each respective tree in the candate set}
}
}
\usage{
phydata
}
\description{
phydose example dataset of two trees
}
\keyword{datasets}
