% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phydose-documentData.r
\docType{data}
\name{phydataDFF}
\alias{phydataDFF}
\title{phydose example distinguishing feature families}
\format{
a list of distinguishing feature families each containing a list of distinguishing f
eatures comprised of vectors of featurettes
}
\usage{
phydataDFF
}
\description{
phydose example distinguishing feature families
}
\keyword{datasets}
