% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phydose-mainfunctions.R
\name{phydose}
\alias{phydose}
\title{Determine the number of cells to sequence in a follow-up single cell experiment}
\usage{
phydose(
  trees,
  fmatrices = NULL,
  umatrices = NULL,
  distFeat = NULL,
  gamma = 0.95,
  fnr = 0,
  kstar_quant = 1
)
}
\arguments{
\item{trees}{a list of binary matrices representing trees with clones as the rows and mutations as the columns}

\item{distFeat}{a optional list of distinguishing features in the same order as the trees}

\item{gamma}{a number that determines the confidence level of the users of observing the tree during sequencing, defaults to 0.95}

\item{fnr}{a number between 0 and 1 that represents the false negative rate of the sequencing technology, defaults to 0}

\item{kstar_quant}{the quantile that determines}

\item{fmatries}{a list of frequency matrices with samples as the rows and mutation frequencies (VAFs) as the columns}
}
\value{
a list that contains the value of k* and a dataframe that contains the value of k for for each tree and sample
}
\description{
Determine the number of cells to sequence in a follow-up single cell experiment
}
\examples{
phydose(trees = phydata$trees, fmatrices = phydata$fmatrices)
phydose(trees = phydata$trees umatrices = phdata$u_list)
\dontrun{
phydose(trees = AML38$trees, fmatrices = AML38$fmatrices, fnr= 0.05)
}
}
