% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phydose-readdata.r
\name{ReadJoint}
\alias{ReadJoint}
\title{Read a file containing a candidate set of trees and an accompnaying F matrix for each tree.}
\usage{
ReadJoint(fname, fmultiplier = 1)
}
\arguments{
\item{fname}{the path to the file that is to be read}
}
\value{
a list
}
\description{
Read a file containing a candidate set of trees and an accompnaying F matrix for each tree.
}
