% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phydose-readdata.r
\name{ReadJointDir}
\alias{ReadJointDir}
\title{Read in data from a directory}
\usage{
ReadJointDir(dir, fmultiplier = 1, fvalues = "fplus", include_u = F)
}
\arguments{
\item{dir}{the directory where the files are stored}
}
\value{
a list
}
\description{
Read in data from a directory
}
