% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phydose-readdata.r
\name{ReadTrees}
\alias{ReadTrees}
\title{Read in set of candidate trees in the given directory and 
convert to a binary matrix with clones as the rows and mutations as columns}
\usage{
ReadTrees(fname)
}
\arguments{
\item{fname}{the path to directory where the trees are located}
}
\value{
A list of binary matrices representing all trees located in the directory
}
\description{
Read in set of candidate trees in the given directory and 
convert to a binary matrix with clones as the rows and mutations as columns
}
