% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phydose_reconcile.r
\name{Reconcile}
\alias{Reconcile}
\title{Calculate the support for each distinguishing feature in a given set of single-cells}
\usage{
Reconcile(dffs, cells)
}
\arguments{
\item{dffs}{a list of distinguishing feature families for trees in the set}

\item{a}{dataframe of sampled single cells}
}
\value{
a dataframe that returns the support for each tree in the set of input cells
}
\description{
Calculate the support for each distinguishing feature in a given set of single-cells
}
