% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phydose-lpconf.R
\name{solveF}
\alias{solveF}
\title{Find the frequencies to use for the k star confidence interval}
\usage{
solveF(tree, fminus, fplus, dff, maxObj = T)
}
\arguments{
\item{tree}{a binary tree matrix}

\item{fminus}{the lower bound of the frequency confidence intervals}

\item{fplus}{the upper bound of the frequency confidence intervals}

\item{dff}{the distinguishing feature of the tree}

\item{maxObj}{a bolean determining if the objective should be maximized for minimized}
}
\value{
a list of lists of featurettes.
}
\description{
For each tree in the input list, generateDistFeat finds the corresponding distinguishing feature family which
is a set of distinguishing features made up of root to mutation paths (featurettes) that unqiuely distinguishes that tree from
all other trees in the input list.
}
