% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phydose-readdata.r
\name{.CreateF}
\alias{.CreateF}
\title{Creates an F matrix from character vector}
\usage{
.CreateF(vecs)
}
\arguments{
\item{vecs}{the character vector from which the frequency matrix is created}
}
\value{
a matrix of the frequencies
}
\description{
Creates an F matrix from character vector
}
