% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phydose-plotting.r
\name{plotTree}
\alias{plotTree}
\title{Plot a tree in the set}
\usage{
plotTree(tree, distfeat = NULL, featurette = NULL, colors = NULL)
}
\arguments{
\item{tree}{an input binary tree}

\item{featurette}{a vector of mutations}

\item{colors}{a vector of colors the length of distFeats or of length 1}

\item{distFeat}{a list of featurettes}
}
\value{
a DiagrammeR graph
}
\description{
Plot a tree in the set
}
